refactor(self-order): extract useOrderingSession hook (Phase 5.1 prep)
Relocate App.tsx's inline state/API logic into a shared hook so upcoming
tablet/landscape/portrait kiosk layouts can reuse it instead of
duplicating menu/cart/order/session logic four times. MobileQRLayout.tsx
now consumes the hook; behavior verified identical (build/typecheck/lint
all clean). App.tsx is a thin shell (always renders MobileQRLayout for
now -- device-type layout selection lands in the next commit).
